Added get_results_page_df() method to the application
This method returns a pandas data frame of the hltv results page. Next possible step is to make a scrapper to initialise database i.e. backfill and another daily one that keeps the data up to date. We may need to make another script that goes in to the match url of the game and collects more information on player performance or if it is a best of 3 collect data on the maps scores not included on the results page